+++ packages/cdk/lambda/assistantHandler/index.ts
@@ -1,56 +1,15 @@
 import { APIGatewayProxyEvent, APIGatewayProxyResult } from 'aws-lambda';
 import {
-  createAssistant,
-  listAssistants,
-  getAssistant,
-  updateAssistant,
-  deleteAssistant,
-  updateKnowledgeSourceStatus,
-  updateAssistantSyncStatus,
-} from '../repository/assistant';
-import { deleteAllMessagesForAssistant } from '../repository/chat';
-import {
-  loadDocuments,
-  chunkDocuments,
-  addMetadata,
-} from '../utils/documentLoader';
-import {
-  indexDocuments,
-  deleteAssistantDocuments,
-} from '../repository/assistantSearch';
-import {
-  Assistant,
-  CreateAssistantRequest,
-  UpdateAssistantRequest,
-  ListAssistantsResponse,
-} from 'generative-ai-use-cases';
-import { getTenantId } from '../utils/tenantUtils';
-import { canAccessAssistant } from '../utils/assistantAccessControl';
-import {
   badRequest400Response,
-  created201Response,
-  forbidden403Response,
   internalServerError500Response,
   methodNotAllowed405Response,
-  noContent204Response,
-  notFound404Response,
-  ok200Response,
 } from '../utils/apiResponse';
-
-/**
- * Helper function to normalize assistant data for API responses
- * - Strips "assistant#" prefix from assistantId
- * - Strips "user#" prefix from userId and id for anonymity and frontend compatibility
- * Internal storage uses prefixed format, but API returns clean values
- */
-function stripAssistantPrefix(assistant: Assistant): Assistant {
-  return {
-    ...assistant,
-    assistantId: assistant.assistantId.replace(/^(assistant#)+/, ''),
-    userId: assistant.userId.replace(/^user#/, ''),
-    id: assistant.id.replace(/^user#/, ''), // Normalize partition key duplicate
-  };
-}
+import { handleCreate } from './create';
+import { handleDelete } from './delete';
+import { handleGet } from './get';
+import { handleList } from './list';
+import { handleUpdate } from './update';
+import * as console from 'node:console';
 
 /**
  * Consolidated handler for all assistant CRUD operations
@@ -102,394 +61,3 @@ export const handler = async (
     return internalServerError500Response({ message: 'Internal Server Error' });
   }
 };
-
-/**
- * Handle POST / - Create assistant
- */
-async function handleCreate(
-  userId: string,
-  event: APIGatewayProxyEvent
-): Promise<APIGatewayProxyResult> {
-  const body: CreateAssistantRequest = JSON.parse(event.body || '{}');
-
-  console.log(
-    `Creating assistant: ragEnabled=${body.ragEnabled}, knowledgeSources=${body.knowledgeSources?.length || 0}`
-  );
-
-  // Basic validation
-  if (!body.name || !body.instruction || !body.modelId) {
-    return badRequest400Response({
-      message: 'Missing required fields: name, instruction, modelId',
-    });
-  }
-
-  const assistant = await createAssistant(userId, body, event);
-
-  // If RAG is enabled, process knowledge sources and update status
-  if (body.ragEnabled) {
-    console.log(`RAG is enabled, checking knowledge sources...`);
-    // If knowledge sources are provided, ingest documents
-    if (body.knowledgeSources && body.knowledgeSources.length > 0) {
-      const cleanAssistantId = assistant.assistantId.replace('assistant#', '');
-
-      // Process each knowledge source individually to track status per-source
-      for (const source of body.knowledgeSources) {
-        try {
-          // Generate ID server-side if not provided (for backward compatibility and URL sources)
-          if (!source.id) {
-            source.id = crypto.randomUUID();
-            console.log(
-              `Generated ID ${source.id} for knowledge source without ID`
-            );
-          }
-
-          console.log(
-            `Processing knowledge source ${source.id} (type=${source.type}, storageKey=${source.storageKey}) for assistant ${cleanAssistantId}`
-          );
-
-          // Update status to SYNCING
-          await updateKnowledgeSourceStatus(
-            assistant,
-            source.id,
-            'SYNCING',
-            undefined,
-            event
-          );
-
-          // Load document for this source
-          const documents = await loadDocuments([source], userId, event);
-
-          // Chunk documents
-          const chunks = await chunkDocuments(documents, 1000, 200);
-
-          // Add metadata
-          const docsWithMetadata = addMetadata(
-            chunks,
-            cleanAssistantId,
-            userId
-          );
-
-          // Index to OpenSearch
-          await indexDocuments(cleanAssistantId, docsWithMetadata, event);
-
-          // Update status to SUCCEEDED
-          await updateKnowledgeSourceStatus(
-            assistant,
-            source.id,
-            'SUCCEEDED',
-            undefined,
-            event
-          );
-
-          console.log(
-            `Successfully ingested knowledge source ${source.id} for assistant ${cleanAssistantId}`
-          );
-        } catch (error) {
-          console.error(
-            `Error ingesting knowledge source ${source.id}:`,
-            error
-          );
-
-          // Update status to FAILED with detailed error message
-          let errorMessage = 'Unknown error';
-          if (error instanceof Error) {
-            errorMessage = error.message;
-
-            // Extract additional details from OpenSearch ResponseError
-            if ('meta' in error && error.meta) {
-              const meta = error.meta as any;
-              if (meta.statusCode) {
-                errorMessage = `${error.message} (HTTP ${meta.statusCode})`;
-              }
-              if (meta.body && meta.body.Message) {
-                // AWS IAM error message format
-                errorMessage += `: ${meta.body.Message}`;
-              } else if (meta.body && meta.body.error) {
-                // OpenSearch error format
-                if (typeof meta.body.error === 'string') {
-                  errorMessage += `: ${meta.body.error}`;
-                } else if (meta.body.error.reason) {
-                  errorMessage += `: ${meta.body.error.reason}`;
-                }
-              }
-            }
-          }
-
-          if (source.id) {
-            await updateKnowledgeSourceStatus(
-              assistant,
-              source.id,
-              'FAILED',
-              errorMessage,
-              event
-            ).catch((statusError) => {
-              // Don't fail if status update fails
-              console.error('Failed to update source status:', statusError);
-            });
-          }
-
-          // Don't fail the assistant creation if one source fails
-          // Continue processing other sources
-        }
-      }
-    } else {
-      console.log(
-        `No knowledge sources provided (knowledgeSources=${body.knowledgeSources?.length || 0})`
-      );
-    }
-
-    // After all sources are processed (or if no sources), update overall assistant status
-    await updateAssistantSyncStatus(assistant, event);
-  } else {
-    console.log(`RAG is not enabled`);
-  }
-
-  // Note: updateAssistantSyncStatus updates assistant.syncStatus in memory
-  return created201Response(stripAssistantPrefix(assistant));
-}
-
-/**
- * Handle GET / - List assistants
- */
-async function handleList(
-  userId: string,
-  event: APIGatewayProxyEvent
-): Promise<APIGatewayProxyResult> {
-  // Read nextToken parameter (aligned with frontend API contract)
-  const nextToken = event.queryStringParameters?.nextToken;
-
-  // Parse and validate limit parameter
-  let limit = 100; // default
-  if (event.queryStringParameters?.limit) {
-    const parsedLimit = parseInt(event.queryStringParameters.limit, 10);
-    if (isNaN(parsedLimit) || parsedLimit < 1 || parsedLimit > 100) {
-      return badRequest400Response({
-        message: 'Invalid limit parameter. Must be between 1 and 100.',
-      });
-    }
-    limit = parsedLimit;
-  }
-
-  try {
-    const result = await listAssistants(userId, event, nextToken, limit);
-
-    // Strip prefix from all assistants
-    // Provide both lastEvaluatedKey (backward compatibility) and nextToken (new standard)
-    const sanitizedResult: ListAssistantsResponse = {
-      assistants: result.assistants.map(stripAssistantPrefix),
-      lastEvaluatedKey: result.lastEvaluatedKey,
-      nextToken: result.lastEvaluatedKey,
-    };
-
-    return ok200Response(sanitizedResult);
-  } catch (error: any) {
-    if (error.message === 'Invalid pagination token') {
-      return badRequest400Response({
-        message: 'Invalid pagination token. Please start from the beginning.',
-      });
-    }
-    throw error;
-  }
-}
-/**
- * Handle GET /{assistantId} - Get assistant
- */
-async function handleGet(
-  userId: string,
-  assistantId: string,
-  event: APIGatewayProxyEvent
-): Promise<APIGatewayProxyResult> {
-  const assistant = await getAssistant(assistantId, event);
-
-  if (!assistant) {
-    return notFound404Response({ message: 'Assistant not found' });
-  }
-
-  // Check access: owner OR (public AND same tenant)
-  if (!canAccessAssistant(assistant, userId, event)) {
-    return forbidden403Response({
-      message: 'Access denied to this assistant',
-      code: 'ASSISTANT_ACCESS_DENIED',
-    });
-  }
-
-  return ok200Response(stripAssistantPrefix(assistant));
-}
-
-/**
- * Handle PUT /{assistantId} - Update assistant
- */
-async function handleUpdate(
-  userId: string,
-  assistantId: string,
-  event: APIGatewayProxyEvent
-): Promise<APIGatewayProxyResult> {
-  const body: UpdateAssistantRequest = JSON.parse(event.body || '{}');
-
-  try {
-    const assistant = await updateAssistant(assistantId, userId, body, event);
-
-    // If knowledge sources were updated and RAG is enabled, re-index documents
-    if (body.knowledgeSources !== undefined && assistant.ragEnabled) {
-      console.log(`Re-indexing documents for assistant ${assistantId}`);
-
-      // KNOWN LIMITATION: We delete old documents before indexing new ones.
-      // If ALL sources fail to index, the assistant will have no documents.
-      // Proper solutions would require:
-      // 1. Adding sync timestamps to documents and deleting only older versions
-      // 2. Implementing async job queue with rollback capability
-      // 3. Using temporary index with atomic swap
-      await deleteAssistantDocuments(assistantId, event);
-
-      // If new knowledge sources are provided, index them
-      if (body.knowledgeSources && body.knowledgeSources.length > 0) {
-        let hasAnySuccess = false;
-        let lastError: Error | undefined;
-
-        // Process each knowledge source individually to track status per-source
-        for (const source of body.knowledgeSources) {
-          try {
-            // Generate ID server-side if not provided (for backward compatibility and URL sources)
-            if (!source.id) {
-              source.id = crypto.randomUUID();
-              console.log(
-                `Generated ID ${source.id} for knowledge source without ID`
-              );
-            }
-
-            console.log(
-              `Processing knowledge source ${source.id} for assistant ${assistantId}`
-            );
-
-            // Update status to SYNCING
-            await updateKnowledgeSourceStatus(
-              assistant,
-              source.id,
-              'SYNCING',
-              undefined,
-              event
-            );
-
-            // Load document for this source
-            const documents = await loadDocuments([source], userId, event);
-
-            // Chunk documents
-            const chunks = await chunkDocuments(documents, 1000, 200);
-
-            // Add metadata
-            const docsWithMetadata = addMetadata(chunks, assistantId, userId);
-
-            // Index to OpenSearch
-            await indexDocuments(assistantId, docsWithMetadata, event);
-
-            // Update status to SUCCEEDED
-            await updateKnowledgeSourceStatus(
-              assistant,
-              source.id,
-              'SUCCEEDED',
-              undefined,
-              event
-            );
-
-            hasAnySuccess = true;
-            console.log(
-              `Successfully re-indexed knowledge source ${source.id} for assistant ${assistantId}`
-            );
-          } catch (error) {
-            console.error(
-              `Error re-indexing knowledge source ${source.id}:`,
-              error
-            );
-
-            // Update status to FAILED with error message
-            const errorMessage =
-              error instanceof Error ? error.message : 'Unknown error';
-            // source.id is guaranteed to exist at this point (generated above if not provided)
-            await updateKnowledgeSourceStatus(
-              assistant,
-              source.id!,
-              'FAILED',
-              errorMessage,
-              event
-            ).catch((statusError) => {
-              // Don't fail if status update fails
-              console.error('Failed to update source status:', statusError);
-            });
-
-            lastError =
-              error instanceof Error ? error : new Error('Unknown error');
-          }
-        }
-
-        // After all sources are processed, update overall assistant status
-        await updateAssistantSyncStatus(assistant, event);
-
-        // If all sources failed, throw error to surface to user
-        if (!hasAnySuccess && lastError) {
-          throw new Error(
-            `Failed to re-index all knowledge sources. Last error: ${lastError.message}`
-          );
-        }
-      } else {
-        console.log(
-          `Cleared all documents for assistant ${assistantId} (no new sources)`
-        );
-
-        // Update status even when clearing all documents
-        await updateAssistantSyncStatus(assistant, event);
-      }
-    }
-
-    return ok200Response(stripAssistantPrefix(assistant));
-  } catch (error: any) {
-    if (error.message === 'Assistant not found') {
-      return notFound404Response({ message: 'Assistant not found' });
-    }
-    if (error.message === 'Unauthorized') {
-      return forbidden403Response({
-        message: 'Access denied to this assistant',
-        code: 'ASSISTANT_ACCESS_DENIED',
-      });
-    }
-    throw error;
-  }
-}
-
-/**
- * Handle DELETE /{assistantId} - Delete assistant
- */
-async function handleDelete(
-  userId: string,
-  assistantId: string,
-  event: APIGatewayProxyEvent
-): Promise<APIGatewayProxyResult> {
-  try {
-    // Delete assistant first (verifies ownership)
-    await deleteAssistant(assistantId, userId, event);
-
-    // Delete all messages after ownership verification
-    await deleteAllMessagesForAssistant(assistantId, event);
-
-    // Delete all indexed documents from OpenSearch
-    try {
-      await deleteAssistantDocuments(assistantId, event);
-      console.log(`Deleted OpenSearch documents for assistant ${assistantId}`);
-    } catch (error) {
-      console.error('Error deleting OpenSearch documents:', error);
-      // Don't fail the deletion if OpenSearch cleanup fails
-    }
-
-    return noContent204Response();
-  } catch (error: any) {
-    if (error.message === 'Assistant not found') {
-      return notFound404Response({ message: 'Assistant not found' });
-    }
-    if (error.message === 'Unauthorized') {
-      return forbidden403Response({
-        message: 'Access denied to this assistant',
-        code: 'ASSISTANT_ACCESS_DENIED',
-      });
-    }
-    throw error;
-  }
-}
